requirements.txt
streamlit==1.28.0
torch==2.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
faiss-cpu==1.7.4
transformers==4.34.0
Pillow==10.0.0
moviepy==1.0.3
google-generativeai==0.3.0
protobuf==3.20.0
safetensors==0.4.0
requests==2.31.0
typing-extensions==4.8.0
packaging==23.2
python-multipart==0.0.6
tqdm==4.66.1
pyyaml==6.0.1
regex==2023.10.3
ftfy==6.1.1
proglog==0.1.10
decorator==4.4.2
imageio==2.31.5
imageio-ffmpeg==0.4.9
